separate and refactor types
