+++ src/addons/Collisions/index.ts
@@ -1,0 +1,1 @@
+export {Collisions} from './Collisions';
